address daniel's concerns about back compat
